+++ protected/humhub/modules/admin/views/space/settings.php
@@ -57,6 +57,8 @@ $this->registerJsConfig('admin.space', [
 
 <?= $form->field($model, 'defaultHideFollowers')->checkbox() ?>
 
+<?= $form->field($model, 'defaultShowTopicSidebar')->checkbox() ?>
+
 <?= $form->field($model, 'allowSpaceTopics')->checkbox(['data' => ['action-change' => 'admin.space.restrictTopicCreation']]) ?>
 
 <?= Button::primary(Yii::t('base', 'Save'))->submit(); ?>